Commitando camada service e ajustes em projeto

+++ src/main/java/br/com/imobiliaria/repository/ImovelRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImovelRepository extends JpaRepository<Imovel, Integer > {
+public interface ImovelRepository extends JpaRepository<Imovel, Long > {
 
 }
